bugfix.  float->double.
git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@13563 16731396-06f5-0310-8873-f7f720988828

+++ public/dataclasses/I3MCWeightDict.h
@@ -27,7 +27,8 @@
  * accessing elements in the map is inherited from the MapPolicy,
  * which is currently set to STLMapStoragePolicy.
  */
-class I3MCWeightDict : public TObject, public STLMapStoragePolicy<string, float> {
+class I3MCWeightDict : public TObject, public STLMapStoragePolicy<string, double> 
+{
  public:
   /**
    * constructor
